+++ config/application.rb
@@ -59,9 +59,6 @@ module Roda
     }
     config.action_mailer.default_url_options = {host: ENV["DOMAIN"]}
 
-    config.skylight.environments += ["pentest", "staging"]
-    config.skylight.probes << "active_job"
-
     # serve dynamic error pages
     config.exceptions_app = routes
   end