+++ terraform/variables.tf
@@ -48,22 +48,6 @@ variable "rollbar_access_token" {
   description = "Rollbar access token"
 }
 
-variable "skylight_access_token" {
-  type        = string
-  description = "Skylight access token"
-}
-
-variable "skylight_env" {
-  type        = string
-  description = "Skylight environment name"
-}
-
-variable "skylight_enable_sidekiq" {
-  type        = string
-  description = "Use Skylight to monitor Sidekiq, true/false"
-  default     = "false"
-}
-
 variable "docker_image" {
   type        = string
   description = "docker image to use"